Fix compare_slave_parameters exit code

+++ src/adapter.h
@@ -62,7 +62,7 @@ typedef struct struct_adapter
 	int dmx_source;
 	int master_source;
 	int is_fbc;
-	int used;
+	uint64_t used;
 	int strength, ber, snr;					   // strength and snr have values between 0 and 255
 	float strength_multiplier, snr_multiplier; // final value: strength * strength_multipler, same for snr
 	uint32_t pid_err, dec_err;				   // detect pids received but not part of any stream, decrypt errors
